+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Re-exported secrecy and bytes types as they are part of the public API.
+
 ## [0.2.0] - 2026-01-04
 
 ### Added